Criação de Dashboard-financeiro com aplicação de gaficos chart.js

+++ front-end/src/app/avaliacao/avaliacao.component.ts
@@ -1,6 +1,8 @@
+import Swal from 'sweetalert2';
 import { Component, OnInit } from '@angular/core';
 import { CookieService } from 'ngx-cookie-service';
 import { Router } from '@angular/router';
+import { AuthService } from '../auth';
 
 @Component({
   selector: 'app-avaliacao',
@@ -8,14 +10,30 @@ import { Router } from '@angular/router';
   styleUrls: ['./avaliacao.component.css']
 })
 export class AvaliacaoComponent implements OnInit {
+  is_superuser: boolean;
 
-  constructor(private cookieService:CookieService,private router: Router ) { }
+  constructor(private cookieService:CookieService,private router: Router, private authServive: AuthService ) { }
 
   ngOnInit(): void {
+  this.is_superuser = this.authServive.getIsSuperUser();
+
   }
   logout() {
-    this.cookieService.delete('jwt','/','localhost',false,'Lax');
-    this.router.navigate(['/login']);
+    Swal.fire({
+      title: 'Tem certeza que deseja sair ?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sim',
+      cancelButtonText: 'Não'
+    }).then((result) => {
+      if (result.isConfirmed){
+      this.cookieService.delete('jwt','/','localhost',false,'Lax');
+      this.cookieService.delete('jwt','/financeiro','localhost',false,'Lax')
+      localStorage.removeItem('selectedEmpresa')
+      this.router.navigate(['/login']);
+
+      }
+    })
   }
 
   setCurrentUser() {
